Look for ;LAYER_CHANGE to end start gcode
Fix issue #13

Prusa Slicer changed its behavior in 2.7.2 and no longer emits ;AFTER_LAYER_CHANGE for single layer prints. So not this code looks for either one of these comments to denote the end of the start gocde.

If there was something custom in the layer change gcode it would be dropped on old versions of Prusa Slicer. Since this is layer specific it seems safe to omit. The generated gcode does all of the layer setup.

* Look for ";LAYER_CHANGE" or ";START_GCODE_END" as alternative way to end the Start G-code block.
* Updated Readme.md
* Updated test pattern to make clear that resetting the extruder to 0 is not part of the comment block.